pyproject.toml
[project]
name = "seek"
version = "0.1.0"
description = "AI assistant for Hyprland with zenity UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
seek = "seek.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
